requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
openai>=1.0.0
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.2.0
jupyter>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
plotly>=5.15.0
